Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -2,8 +2,8 @@
 int main() {
     long long  n;
     scanf("%lld",&n);
-    int size=sizeof(n)*8;
-    int y=n>>(size-1);
+    long size=sizeof(n)*8;
+    long y=n>>(size-1);
     if(n&1==1){
         printf("Set");
     }
